clusterizer: Cleanup documentation and interface
Instead of referring to the micro-index buffer as "index", we now
consistently refer to it as "triangles"; this is reflected in renaming
meshlet_indices to meshlet_triangles and meshopt_Meshlet::index_offset
to triangle_offset.

This change also swaps the order of arrays in computeMeshletBounds to be
consistent between all functions.