Ensure hash tables always have some percentage of free buckets
While in most cases we overallocate the hash tables as the number of
elements we supply is an upper bound that's never reached, it's possible
to create meshes that do reach that bound by having a power of two count
of vertices and have all vertices be unique.

To avoid catastrophic performance degradation of hash-based algorithms
on such meshes we make sure that we have ~20% of elements free, which
means that our expected chain length is reasonably short in all cases.

Note: this isn't motivated by a specific example, but is merely done as
a precaution.