gltfpack: Implement initial support for KHR_materials_variants
This change implements basic support for material variants. Each mesh
now keeps track of a set of material mappings; mesh merging is limited
to cases when the variant set is the same, and all material variants are
preserved.

Since we don't correctly compute the UV bounds for variant materials,
this only works without quantization so far.