gltfpack: Combine variant quantization bounds
When a mesh uses multiple materials through the use of variants, we need
to make sure that the quantization parameters for UV for a given
material incorporate all meshes that refer to it, including variants.

In addition, when processing a mesh, we need to take all variant
requirements into account, so that e.g. if only one material needs a
second UV set, we still include it even if it's a variant.

This isn't quite correct in cases when materials are used in other
meshes as well, since we need to ensure that the quantization data is
the same across all materials.