gltfpack: Fix texture compression spawn in node.js
Our previous code blindly replaced the executable in the input command
with the env. variable value, which actually didn't handle cases when
the environment variable wasn't set and additionally didn't handle paths
with spaces well.

We now use shell:false execution option and give it a pre-split command
line - this doesn't work for arguments with spaces but we never handled
them correctly (and also they shouldn't occur as the paths that are
passed to the compressor are that of temp files...).