simplify: Fix a significant hash quality issue for integer coordiantes
When the input coordinates to the hash function were integers (in
floating point form), the resulting hash function was incredibly poor.

It generated too many collisions because it mostly used the entropy from
the lower bits, and for integer values many lower bits are 0.

This change fixes the problem by lightly mixing the bits before hashing,
which seems to completely fix this issue in practice. On a Minecraft
export, the stats are as follows:

Total number of vertices (with many duplicates): 13 408 528
Number of unique 32-bit hashes with Murmur: 2 525 651
Number of unique 32-bit hashes with spatial hash: 205 457
Number of unique 32-bit hashes with scrambled spatial hash: 2 525 626