gltfpack: Switch to uastc_rdo_l for UASTC texture compression
This change is necessary because unfortunately latest versions of basisu
and toktx changed the parameter name. The new parameter name has a
different meaning but the table tuned before works reasonably well, so
only the parameter name needs to change.

Since we have to drop support for earlier versions of the tools anyway,
this change goes ahead and also removes .basis -> KTX2 conversion in
favor of basisu's new -ktx2 flag.

This change is still compatible with older versions of toktx when using
ETC1S encoding, but for simplicity requires basisu 1.15 regardless of
the format.