gltfpack: Implement support for normalized position output
This change adds -vpn flag that enables the use of normalized position
accessors instead of integer (unnormalized) ones.

The benefits, similarly to texture coordinates, include improved shader
compatibility for APIs like WebGPU, as well as improved support for
lower precision storage of scale values. In particular, Unity doesn't
support scale values below a certain range, so this change allows
producing meshes that are fully compatible with Unity.

Unfortunately, this mode also has downsides. Because we need to use
signed values for morph target deltas, it requires us to adjust the
target deltas to match position scale which in effect reduces effective
precision on deltas by one bit, as well as prohibiting the use of 8-bit
target deltas. Additionally, this change requires renderers to rescale
the positions on CPU for features like ray casting etc so models
compressed with -vpn are likely not going to work in Three.JS.