indexgenerator: Allow unindexed generateVertexRemap for point clouds
generateVertexRemap was originally intended for triangle meshes, but it
can work for any set of data, and can be useful to reindex vertex/point
data separately. For now let's specifically allow the use case of
calling it without an index buffer on an arbitrary data set, and keep
the mod 3 restriction for indexed data to keep flagging accidental
incorrect usage.