gltfpack: Switch to basis_parallel_compress
When internal Basis encoding was implemented initially, Basis didn't
provide an API to compress multiple textures together in parallel, but
now it does.

Right now we rely on the gltfpack fork to provide a special performant
job pool implementation that allows us to scale parallel compression
much better. As a side effect, this means gltfpack *must* be built
against the forked version of basis_universal, which results in
confusion around builds.

The implementation of parallel_compress is inefficient in the main
version, but it's efficient in gltfpack fork. This change switches to
that API instead, with the eventual angle of integrating the efficiency
improvements into mainline so that we no longer need to maintain a fork.

Before this change we would read the images in parallel, but that didn't
really provide appreciable savings so the resulting change is neutral on
performance.

Building against the upstream version now works, and simply results in
less efficient texture compression.