gltfpack: Experimental support for floating point position quantization
Currently gltfpack provides all or nothing quantization: either all
attributes get quantized, or none do.

When quantization is used, we use the most size and memory efficient
formats, which is great, but it can lead to integration issues for
applications - since dequantization uses an extra transform, this
results in changes that are needed for some applications to take that
transform into account. Notably, these changes mostly revolve around
positions (and rarely around texture coordinates wrt texture
replacement, but positions are definitely the main source of issues).

Due to the flexibility of meshopt codecs however, there are other
options available that don't change the coordinate space of mesh
positions, but instead just change the positions a little bit to encode
more efficiently.

This change adds a new mode, -vpf, which uses floating point
quantization to encode position values:

- When using regular compression (-c), we quantize each coordinate
  individually
- When using extra compression (-cc), we use exponential filter when
  serializing the position stream

It is expected that either mode is less efficient than integer
quantization, but -cc can be closer to integer quantization and on
meshes with many attributes the extra size could be an acceptable
tradeoff for better compatibility.