gltfpack: Fix position accessor bounds when using -vpf
We need to quantize the bounds with the same algorithm we apply to
vertex data.

Technically, when using meshopt compression, we also need to apply
exponent encoding to bounds. However, in practice the delta is small and
the validator can't decode meshopt data, so we'll leave this until
another day.