vertexcodec: Implement scalar compare+popcnt for SSE3/NEON
This redundantly recomputes the number of sentinel elements in the
source byte group using scalar arithmetics exclusively; the general
strategy is to put 0 or 1 into each nibble, and then add them all into
the top nibble using a 64-bit multiply + shift. The sum doesn't overflow
because mode1/2 would never be used when all elements are sentinel,
since mode3 is always picked in these cases as it's shorter.

The reason why this helps is that the computation is shorter
(latency-wise) than the actual decoding that uses SIMD, and this allows
the CPU to overlap the decoding of consecutive byte groups.

This may regress throughput on in-order CPUs but should be generally
beneficial on CPUs where vector and scalar instructions can co-issue
with a reasonably wide window.

Based on idea suggested by @dougallj