Don't add -Werror on GCC, it can break compilation in certain cases.
The _mm_undefined_si128() AVX-512 intrinsics on GCC produces an
"uninitialized variable" warning (which is, well, what is it intended to
actually do), and that makes the code not compile even though there's no
actual error. I discovered this only because I use -march=native to
build system packages and recently upgraded to an AVX-512-equipped
hardware.

I assume this is just a temporary compiler wart that will eventually get
fixed, so there's no point in trying to silence that warning from
within meshoptimizer.