clusterizer: Improve selection critera of first triangle of new meshlet
When a meshlet is full, the triangle we've selected starts the new meshlet.

That triangle has most often been selected using meshlet scoring based on
size and cone - however, these are in relation to the old meshlet that the
new triangle doesn't relate to.

This change replaces the scoring criteria for cases like this with topological
criteria based on number of outstanding triangles in the proximity. We want to
start with triangles that have the smaller number of live neighbors because
otherwise we're more likely to leave small isolated islands of triangles which
ends up increasing the meshlet size.

This change ends up reducing the number of meshlets by 1-2% and reducing the
mean size of a meshlet by up to 5%, yielding slightly more optimal output for
roughly the same build cost (since we only end up re-scoring once every ~90
triangles for 64 vertex meshlets).

Note: technically, "extra" variable computed during scoring function is also
meaningless. However, removing it from the second scoring results in substantial
degradation in meshlet packing efficiency - it looks like this is for similar
reasons, where we really want to prioritize starting triangles that would have
otherwise become dead-ends in other traversals, and not doing this degrades
quality overall even though it may slightly improve selection for individual
meshlets.