js: Fix significant overallocation for simplify
When computing the number of vertices, we took the length of the
Float32Array without accounting for the stride. This would typically
result in 3-4x higher value that's being passed in, which would
significantly increase the memory consumption during simplification, as
the array of quadrics would be much larger than it needed to be.