gltfpack: Enable border locking during simplification if requested
Locking border vertices may be useful to preserve geometric details in
certain cases. This helps when the mesh is split into multiple chunks
(especially if the splitting is based on material since gltfpack won't
be able to merge the sub meshes before simplification), or on certain
classes of 3D scanned meshes when triangles are covering gaps in
geometry without being connected to anything.

This setting is opt-in as it regresses simplification quality in cases
when the locking isn't required; for some applications it's a safer
default though so they can choose to enable -slb always.