vertexcodec: Clean up Wasm SIMD implementation
When Wasm SIMD spec was finalized, the SIMD intrinsics in LLVM got
renamed to match the spec. This change happened in LLVM 13; the old
intrinsics emit deprecation warnings.

Since it's been two years and we're now on LLVM 15 in the official
builds, it's probably time to clean up the code and switch to new
intrinsic names. If we need to we can add stubs for LLVM 12 separately;
this only affects builds that explicitly target Wasm SIMD so that is
unlikely.

Also, we don't need to disable unused variable warning since we can just
remove the variables that were never used to begin with.