gltfpack: Preserve custom ID attributes
This change keeps attributes with _ID_ or _ID in the name in tact when
processing the scene. We assume that these attributes are scalar and
encode them as floating-point (to avoid possible range issues with uint16);
the filter doesn't seem to be necessary here although we could use an
exponential filter for slightly better compression in the future.