gltfpack: Use exp encoding for integer ID attributes
This results in ~2x smaller attribute streams so it is very worthwhile;
as with other compression filters, we only apply this when -cc is
specified.

The encoding assumes that the values are in 0..2^23-1 range; since input
values may be encoded as floats which permits a slightly wider range
(integers up to 2^24 are encoded exactly), we guard against this to
make sure no data loss occurs.