quantization: Implement meshopt_dequantizeHalf
This function reverses the transformation by meshopt_quantizeHalf; the
reverse conversion is easier since exponent and mantissa just need to be
expanded into the larger range. We still need to handle denormals and
specials; unlike quantization, here we choose to preserve NaN payload
because that allows us to fold inf and nan cases into a single
conditional add.