gltfpack: Implement texture copying when compression is not used
When compression is not used and the output is a .gltf file, we currently
simply preserve the image URIs and expect the caller to somehow copy the
images. This is convenient for testing, but results in glTF files that
are missing the image resources.

With this change, the newly added option -tcp uses the flow similar to
the one we use for compression to write out copies of image assets when
requested.