gltfpack: Remove point normal streams when they are constant
A fair amount of point cloud exports to glTF seem to have a redundant
normal stream (filled with the same value). This is relatively common
(about one in 6-7 files from Sketchfab has this issue), and it results
in incorrect rendering for renderers that respect normals *and* extra
data (4 out of 16 bytes in point clouds with normals & colors).

While it may sometimes be incorrect to remove this data, for now let's
err on the side of aggressive optimization and detect & remove it.