simplify: Fix quadricFromPoint math
Before this, point simplifier encoded non-sensical functions into the
quadric which resulted in fairly suboptimal candidate point selection,
raising the effective error (when computed separately as a distance function)
by ~2x.

To make mistakes like this less likely in the future we now annotate
each quadric structure with the formula it encodes.